requirements.txt
Django==1.10.8
coverage==3.6
dj-database-url==0.2.1
gunicorn==0.17.4
psycopg2==2.7
dj-static==0.0.6
django-reversion==1.10.2
django-axes==1.7.0
ffs==0.0.8.2
letter==0.4.1
requests==2.18.4
-e git+https://github.com/srosro/django-test-coverage@3f34cc05e50dd16b75d2e18a681d133f27dc5668#egg=django_test_coverage
djangorestframework==3.4.7
django-compressor==1.5
opal==0.10.0
jsonfield==1.0.3
